feat: TASK_015/016/017 integration — wavetable, granular, sample depth
Engine fixes:
- MidiPipeline: readIntNormAP() now reads raw int values directly;
  old normalised branch misread ratchet/transpose params (ratchet=1→4)
- TheoryEngine: seedDatabase() upgraded to version-stamp system
  (kProgSeedVersion=3); replaces fragile count/root-sequence checks
- WolfsDenParameterLayout: wt_index_a/b range 0-100→0-19 (20 tables)
- SynthEngine: per-block WT dirty check — wavetable switches real-time
  without plugin restart; cachedWtIndexA/B[4] added to private state
- WavetableBank: sign-conversion warnings eliminated (int→size_t)

UI fixes:
- CompositionPage: RATE/PATTERN/OCTAVES/SWING labels added above arp controls
- TopBar: deprecated juce::Font(float) → juce::Font(juce::FontOptions(float))
- SynthPage: unused variable/param warnings eliminated

Build: zero errors, zero warnings — VST3, AU, Standalone all clean

+++ 05_ASSETS/apex_project/Source/engine/synth/WavetableBank.h
@@ -12,21 +12,21 @@ template<typename Func>
 std::array<double, kWavetableSize> makeWavetable(Func f)
 {
     std::array<double, kWavetableSize> table{};
-    for (int i = 0; i < kWavetableSize; ++i)
+    for (size_t i = 0; i < (size_t)kWavetableSize; ++i)
     {
-        double phase = static_cast<double>(i) / kWavetableSize;
+        double phase = static_cast<double>(i) / static_cast<double>(kWavetableSize);
         table[i] = f(phase);
     }
     // Normalize to [-1, 1]
     double peak = 0.0;
-    for (int i = 0; i < kWavetableSize; ++i)
+    for (size_t i = 0; i < (size_t)kWavetableSize; ++i)
     {
         if (std::abs(table[i]) > peak)
             peak = std::abs(table[i]);
     }
     if (peak > 0.0)
     {
-        for (int i = 0; i < kWavetableSize; ++i)
+        for (size_t i = 0; i < (size_t)kWavetableSize; ++i)
             table[i] /= peak;
     }
     return table;
